requirements/base.txt
Django==1.11.2
django-environ==0.4.3
django-extensions==1.7.9
flake8==3.3.0
ipdb==0.10.3
ipython==6.1.0
